增加代码生成器,待完善

+++ code/src/main/java/cn/swallowff/code/gen/GeneratorImpl.java
@@ -1,12 +1,15 @@
 package cn.swallowff.code.gen;
 
 import cn.swallowff.code.FileType;
-import cn.swallowff.code.factory.TemplateDataFactory;
 import cn.swallowff.code.config.GeneratorConfig;
 import cn.swallowff.code.entity.TemplateData;
+import cn.swallowff.code.exception.GenerationException;
+import cn.swallowff.code.exception.InvalidConfigException;
 import cn.swallowff.code.factory.FilePathFactory;
+import cn.swallowff.code.factory.TemplateDataFactory;
 import cn.swallowff.code.resolver.MetaDataResolver;
 import cn.swallowff.code.util.BeetlUtil;
+import cn.swallowff.code.util.PathUtils;
 import cn.swallowff.code.util.StreamUtil;
 import cn.swallowff.common.io.FileUtils;
 import cn.swallowff.common.mapper.BeanMapConvert;
@@ -14,6 +17,7 @@ import org.beetl.core.Template;
 
 import java.io.*;
 import java.nio.charset.Charset;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -30,14 +34,59 @@ public class GeneratorImpl implements IGenerator {
         this.generatorConfig = generatorConfig;
     }
 
-    public void init(){
+    protected void init() throws InvalidConfigException{
+        checkConfig();
         templateData = new TemplateDataFactory(new MetaDataResolver(generatorConfig)).createTemplateData();
         pathMap = new FilePathFactory(generatorConfig).createPathMap();
     }
 
+    //TODO 验证配置文件
+    protected void checkConfig() throws InvalidConfigException {
+        File file = new File(PathUtils.PROJECT_ROOT + File.separator + generatorConfig.getModuleLocation());
+        if (!file.isDirectory()){
+            throw new InvalidConfigException("module location must be a path where exist in your project path , make sure it is a effective directory ");
+        }
+
+    }
+
     @Override
-    public void doGen(){
+    public void execute() throws GenerationException {
         init();
+        List<FileType> genFiles = generatorConfig.getGenFiles();
+        if (genFiles.contains(FileType.ENTITY)){
+
+        }
+        if (genFiles.contains(FileType.DAO)){
+
+        }
+        if (genFiles.contains(FileType.SERVICE)){
+
+        }
+        if (genFiles.contains(FileType.CONTROLLER)){
+
+        }
+        if (genFiles.contains(FileType.MAPPER)){
+
+        }
+        if (genFiles.contains(FileType.LIST_HTML)){
+
+        }
+        if (genFiles.contains(FileType.ADD_HTML)){
+
+        }
+        if (genFiles.contains(FileType.EDIT_HTML)){
+
+        }
+        if (genFiles.contains(FileType.LIST_JS)){
+
+        }
+        if (genFiles.contains(FileType.ADD_JS)){
+
+        }
+        if (genFiles.contains(FileType.EDIT_JS)){
+
+        }
+
         genEntity();
         genDao();
         genService();
@@ -51,68 +100,59 @@ public class GeneratorImpl implements IGenerator {
         genEditJs();
     }
 
-    @Override
-    public void genEntity() {
+    protected void genEntity() {
         writeFile(FileType.ENTITY);
-
     }
 
-    @Override
-    public void genDao() {
+    protected void genDao() {
         writeFile(FileType.DAO);
     }
 
-    @Override
-    public void genService() {
+    protected void genService() {
         writeFile(FileType.SERVICE);
     }
 
-    @Override
-    public void genController() {
+    protected void genController() {
         writeFile(FileType.CONTROLLER);
     }
 
-    @Override
-    public void genXmlMapper() {
+    protected void genXmlMapper() {
         writeFile(FileType.MAPPER);
     }
 
-    @Override
-    public void genListPage() {
+    protected void genListPage() {
         writeFile(FileType.LIST_HTML);
     }
 
-    @Override
-    public void genAddPage() {
+    protected void genAddPage() {
         writeFile(FileType.ADD_HTML);
     }
 
-    @Override
-    public void genEditPage() {
+    protected void genEditPage() {
         writeFile(FileType.EDIT_HTML);
     }
 
-    @Override
-    public void genListJs() {
+    protected void genListJs() {
         writeFile(FileType.LIST_JS);
     }
 
-    @Override
-    public void genAddJs() {
+    protected void genAddJs() {
         writeFile(FileType.ADD_JS);
     }
 
-    @Override
-    public void genEditJs() {
+    protected void genEditJs() {
         writeFile(FileType.EDIT_JS);
     }
 
-    public void writeFile(FileType fileType){
+    protected void writeFile(FileType fileType){
         FileOutputStream fos = null;
         OutputStreamWriter osw = null;
         BufferedWriter bw = null;
         try {
-            FileUtils.createFile(pathMap.get(fileType));
+            boolean created = FileUtils.createFile(pathMap.get(fileType));
+            if (!created && !generatorConfig.isForceCover()){
+                return;
+            }
 //            FileUtils.createDirectory(StringUtils.substringBeforeLast(pathMap.get(fileType),"/"),);
             fos = new FileOutputStream(pathMap.get(fileType));
             osw = new OutputStreamWriter(fos,Charset.forName("utf-8"));
@@ -132,7 +172,7 @@ public class GeneratorImpl implements IGenerator {
         }
     }
 
-    public Template getTemplate(FileType fileType) throws IOException {
+    protected Template getTemplate(FileType fileType) throws IOException {
         return BeetlUtil.getTemplate(fileType.getTemplateName(),generatorConfig.getTemplatePath(),"utf-8");
     }
